testing fix to end exit on cutdown script

+++ builders/deb-build.sh
@@ -244,17 +244,22 @@ cutdown() {
 					mkdir -p $CHROOT_DIR/usr/share/man/man${p}
 				done ;;
 			nls)
+			  if [ -d "$CHROOT_DIR/usr/share/locale" ] &&
+				 [ `ls | grep -c -m 1 '.*'` -eq 1 ]; then 
 				[ -d $CHROOT_DIR/usr/lib64/locale ] && LIBDIR=lib64
 				rm -rf $NLS_DIR; mkdir -p $NLS_DIR/usr/share/locale $NLS_DIR/usr/$LIBDIR/locale
 				for p in $(ls $CHROOT_DIR/usr/share/locale); do
 					[ $p != en ] && mv $CHROOT_DIR/usr/share/locale/$p $NLS_DIR/usr/share/locale
 				done
+			  if [ -d "$CHROOT_DIR/usr/$LIBDIR/locale" ] &&
+				 [ `ls | grep -c -m 1 '.*'` -eq 1 ]; then 
 				for p in $(ls $CHROOT_DIR/usr/$LIBDIR/locale); do #TODO: add check to see if the directory is empty, otherwise the script might unexpectly exit
 					case $p in
 						en_US|en_AU|en_US.*|en_AU.*|C|C.*) ;; # skip
 						*) mv $CHROOT_DIR/usr/$LIBDIR/locale/$p $NLS_DIR/usr/$LIBDIR/locale
 					esac
 				done ;;
+			  fi
 			dev)
 				# recreates dir structure, move headers and static libs to devx dir
 				rm -rf $DEVX_DIR
